Create two separate trace paths for Mac and Windows
Mac needed to get the binary from the packaged info.plist in order to work universally.

Testing is still required.
