+++ go.mod
@@ -3,9 +3,13 @@ module github.com/jpeizer/Vectorworks-Utility
 go 1.16
 
 require (
-	github.com/AllenDang/giu v0.5.4-0.20210426041429-9dee43b7338d // indirect
-	github.com/AllenDang/imgui-go v1.12.1-0.20210421054105-2098c13761b7 // indirect
-	github.com/sqweek/dialog v0.0.0-20200911184034-8a3d98e8211d // indirect
+	github.com/AllenDang/giu v0.5.4-0.20210426041429-9dee43b7338d
+	github.com/AllenDang/imgui-go v1.12.1-0.20210421054105-2098c13761b7
+	github.com/TheTitanrain/w32 v0.0.0-20200114052255-2654d97dbd3d // indirect
+	github.com/go-gl/gl v0.0.0-20210426225639-a3bfa832c8aa // indirect
+	github.com/go-resty/resty/v2 v2.6.0 // indirect
+	github.com/sqweek/dialog v0.0.0-20200911184034-8a3d98e8211d
+	golang.org/x/net v0.0.0-20210428140749-89ef3d95e781 // indirect
 	golang.org/x/sys v0.0.0-20210426230700-d19ff857e887
 	howett.net/plist v0.0.0-20201203080718-1454fab16a06
 )
