+++ internal/ui/traceApplication.go
@@ -1,14 +1,11 @@
 package ui
 
 import (
-	"bufio"
 	"bytes"
 	g "github.com/AllenDang/giu"
 	"github.com/AllenDang/imgui-go"
 	"github.com/sqweek/dialog"
-	"io"
 	"log"
-	"os/exec"
 )
 
 var buffer bytes.Buffer
@@ -44,38 +41,4 @@ func RenderTraceApplication() g.Widget {
 			imgui.Button("Submit")
 		}
 	})
-}
-
-func runApplication(ch chan []byte, targetFile string) {
-	cmd := exec.Command(targetFile)
-
-	outReader, outWriter := io.Pipe()
-	errReader, errWriter := io.Pipe()
-
-	cmd.Stdout = outWriter
-	cmd.Stderr = errWriter
-
-	//mw := io.MultiWriter(outWriter, errWriter, &buffer)
-	mr := io.MultiReader(outReader, errReader, &buffer)
-
-
-
-	// Is the transfer the output of the application, through the channel.
-
-	go func() {
-		reader := bufio.NewReader(mr)
-
-		for {
-			line, _, err := reader.ReadLine()
-			if err != nil {
-				break
-			}
-			ch <- line
-			ch <- []byte{'\n'}
-		}
-	}()
-
-	if err := cmd.Run(); err != nil {
-		log.Panicln(err)
-	}
 }
